requirements.txt
requests>=2.31.0
jinja2>=3.1.2
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2
backoff>=2.2.1
aiohttp>=3.9.1
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
jupyter>=1.0.0
